change: Rename EntityInfo module and convert it to TypeScript
Use the full name of the component also for the module and add types.
